build: fix smartcast dependnecy
As mentioned in the PR for smartcast it seems like a bug in the mps build
script generate that doesn't emit the code the extract a zip file if it is
used as a dependency. I changed smartcasts layout to only contain the plugin
wihtout packaging it as a zip.

Here I changes the build file to consume smartcasts.